ai-service-python/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
google-generativeai==0.3.2
transformers==4.35.0
torch==2.1.0
sentencepiece==0.1.99
numpy<2.0.0
